r082 Refactor model and viewmodel
Refactor the model in line with MVVM principles: The model is now wholly independent of view and viewmodel structures, but does little more than fetch and save data from the server. There's no real business logic contained within the model.

The viewmodel contains most of the real business logic and knowledge of the model...