Update the event manager to point to specific stream endpoints
- /broadcast/stream   is used for individual updates (i.e. each time new data is entered)
- /broadcast/result   is used for updates to the route result

A single @url value is used for the broadcaster. This is currently set to 10.0.2.10 (as the NGINX server is on a separate box), it will likely need changing to 0.0.0.0 when the application is hosted on the server. It may also need to be amended during testing is the IP address assigned to melody is changed.