Update accessors_results
Update the locked method.
Instead of retrieving the results dataset and performing a series of updates, create a temporary table and update using a cross join on the old and new data
This is about 5-6 * faster than the iterative method, or 3-4 * faster if the iterative update is wrapped in a DB.transaction loop.